Split systemDetailSections.tsx into evidence/bodies-stations/shell files
systemDetailSections.tsx was 720 lines, ~2.5x the next-largest file in
its directory (systemDetailPlannerEntry.tsx at 291). Item 6 of the
ongoing code-splitting refactor series (item 1: PlannerCanvasPreview.tsx
removal, PR #419; item 2: api.ts domain-module split, PR #420; item 3:
MyWorkWorkspace.tsx split, PR #421; item 4: R3FMapFoundation.tsx
scene-concern split, PR #422; item 5: SearchForm.tsx split, PR #423).

Pure code movement, no logic changes. Follows the same compatibility-
barrel pattern as the api.ts split: SystemDetailModal.tsx's import
from './systemDetailSections' is unchanged.

- SystemDetailSectionShell.tsx: the shared Section wrapper primitive,
  used by every other section component.
- SystemEvidenceSection.tsx: SystemEvidenceSection and its private
  evidence-lifecycle helpers.
- SystemBodiesAndStationsSections.tsx: BodiesSection and StationsSection
  and their private table/badge helpers.
- systemDetailSections.tsx (720 -> 204 lines): keeps ModalHeader,
  SystemInfoGrid, ExplorationValue, ExternalLinks directly, re-exports
  Section/SystemEvidenceSection/BodiesSection/StationsSection from the
  three new files.

Verified only the 8 originally-exported symbols are imported outside
this file (all from SystemDetailModal.tsx, which needed zero changes).
typecheck, lint (0 errors, same pre-existing warning baseline), knip,
the system-detail-scoped test suite (40 files / 252 tests, incl.
SystemDetailModal.test.tsx's 18), and production build all pass.
