+++ frontend/src/features/system-detail/SystemDetailSectionShell.tsx
@@ -1,0 +1,14 @@
+import type { ReactNode } from 'react';
+
+export function Section({ title, children }: { title: string; children: ReactNode }) {
+  return (
+    <section className="space-y-3">
+      <h3 className="font-mono text-silver-dk uppercase tracking-[0.22em] text-[10px] flex items-center gap-3">
+        <span className="block h-px flex-1 bg-gradient-to-r from-transparent via-border-bright to-transparent" />
+        <span className="text-orange">{title}</span>
+        <span className="block h-px flex-1 bg-gradient-to-r from-border-bright via-border-bright to-transparent" />
+      </h3>
+      {children}
+    </section>
+  );
+}
